# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AlfaUnipac GV — Sistema de Gestão Acadêmica API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /turmas:
    get:
      operationId: listTurmas
      tags: [turmas]
      summary: List all turmas
      security:
        - bearerAuth: []
      parameters:
        - name: disciplinaId
          in: query
          schema:
            type: integer
        - name: periodo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of turmas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Turma"
    post:
      operationId: createTurma
      tags: [turmas]
      summary: Create a turma
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTurmaBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Turma"
  /turmas/{id}:
    get:
      operationId: getTurma
      tags: [turmas]
      summary: Get a turma by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Turma
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TurmaDetalhada"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTurma
      tags: [turmas]
      summary: Update a turma
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTurmaBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Turma"
  /turmas/{id}/alunos:
    get:
      operationId: getTurmaAlunos
      tags: [turmas]
      summary: Get students enrolled in a turma
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of students in turma
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Aluno"
  /turmas/{id}/matricular:
    post:
      operationId: matricularAluno
      tags: [turmas]
      summary: Enroll a student in a turma
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatricularAlunoBody"
      responses:
        "201":
          description: Enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Matricula"

  # CHAMADAS (ATTENDANCE)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Aluno:
      type: object
      properties:
        id:
          type: integer
        nomeCompleto:
          type: string
        cpf:
          type: string
        matricula:
          type: string
        curso:
          type: string
        turno:
          type: string
          enum: [Matutino, Vespertino, Noturno]
        status:
          type: string
          enum: [Ativo, Retido, Trancado, Egresso]
        valorMensalidade:
          type: number
        financiador:
          type: string
        createdAt:
          type: string
      required:
        - id
        - nomeCompleto
        - cpf
        - matricula
        - curso
        - turno
        - status
        - valorMensalidade
        - financiador
        - createdAt
...
    Disciplina:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        curso:
          type: string
        semestre:
          type: integer
        cargaHorariaTotal:
          type: integer
        professor:
          type: ["string", "null"]
      required:
        - id
        - nome
        - curso
        - semestre
        - cargaHorariaTotal
...
    Turma:
      type: object
      properties:
        id:
          type: integer
        disciplinaId:
          type: integer
        periodo:
          type: string
        dataInicio:
          type: string
        dataFim:
          type: string
        disciplinaNome:
          type: ["string", "null"]
        curso:
          type: ["string", "null"]
      required:
        - id
        - disciplinaId
        - periodo
        - dataInicio
        - dataFim
    TurmaDetalhada:
      type: object
      properties:
        id:
          type: integer
        disciplinaId:
          type: integer
        periodo:
          type: string
        dataInicio:
          type: string
        dataFim:
          type: string
        disciplina:
          $ref: "#/components/schemas/Disciplina"
        totalAlunos:
          type: integer
      required:
        - id
        - disciplinaId
        - periodo
        - dataInicio
        - dataFim
        - disciplina
        - totalAlunos
    CreateTurmaBody:
      type: object
      properties:
        disciplinaId:
          type: integer
        periodo:
          type: string
        dataInicio:
          type: string
        dataFim:
          type: string
      required:
        - disciplinaId
        - periodo
        - dataInicio
        - dataFim
    UpdateTurmaBody:
      type: object
      properties:
        disciplinaId:
          type: integer
        periodo:
          type: string
        dataInicio:
          type: string
        dataFim:
          type: string
    Matricula:
      type: object
      properties:
        id:
          type: integer
        turmaId:
          type: integer
        alunoId:
          type: integer
      required:
        - id
        - turmaId
        - alunoId
    MatricularAlunoBody:
      type: object
      properties:
        alunoId:
          type: integer
      required:
        - alunoId

    # Chamada (Attendance)
...